requirements.txt
annotated-types==0.7.0
anyio==4.12.1
arcadepy==1.10.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
claude-agent-sdk==0.1.25
click==8.3.1
cryptography==46.0.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.26.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.22
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
openai>=1.0.0
httpx>=0.25.0
google-genai>=1.0.0
groq>=0.11.0
